Switch from a pod to a deployment.
Gravity 7 has new pod security policies that can be incorrectly applied
to previously working user applications as discussed at
https://github.com/gravitational/gravity/issues/2105.

Thus we need to specify a security context for the alpine pod.  However,
helm upgrade (but not install) fails when a explicit security context is
set on a pod.  The recommended way to migrate this is a deployment --
which will successfully install or upgrade with securityContext set.