Add cluster manifests for gravity upgrade training.
I wanted to have a cluster that had an application with upgrade hooks
defined, such that there would be something visible that changes during
the user app phase.

I chose 7.0.12 as the first LTS release on 7.x (some of the early 7.x
had some upgrade stability issues) and 7.0.30 as the most
current 7.x at the time of writing. Unfortunately, this doesn't include
a planet etcd upgrade, but it covers most other interesting parts of the
upgrade.